dbt_dagster_project/dbt_dagster_project/assets.py
from dagster import AssetExecutionContext
from dagster_dbt import DbtCliResource, dbt_assets

from .project import dbt_demo_project


@dbt_assets(manifest=dbt_demo_project.manifest_path)
def dbt_demo_dbt_assets(context: AssetExecutionContext, dbt: DbtCliResource):
    yield from dbt.cli(["build"], context=context).stream()
    